move addNewObserver to zetae2e
