+++ cmd/zetae2e/local/tss_migration.go
@@ -25,7 +25,7 @@ func tssMigrationTestRoutine(
 		account := conf.AdditionalAccounts.UserMigration
 		// initialize runner for migration test
 		tssMigrationTestRunner, err := initTestRunner(
-			"tssMigration",
+			"triggerTSSMigration",
 			conf,
 			deployerRunner,
 			account,
@@ -65,7 +65,7 @@ func tssMigrationTestRoutine(
 	}
 }
 
-func TSSMigration(deployerRunner *runner.E2ERunner, logger *runner.Logger, verbose bool, conf config.Config) {
+func triggerTSSMigration(deployerRunner *runner.E2ERunner, logger *runner.Logger, verbose bool, conf config.Config) {
 	migrationStartTime := time.Now()
 	logger.Print("🏁 starting tss migration")
